complete mypage style
