+++ app/src/main/java/com/depromeet/android/childcare/mypage/CategoryChartView.kt
@@ -3,8 +3,15 @@ package com.depromeet.android.childcare.mypage
 import android.content.Context
 import android.graphics.Color
 import android.util.AttributeSet
+import androidx.core.content.ContextCompat
+import com.depromeet.android.childcare.R
 import com.github.mikephil.charting.charts.HorizontalBarChart
 import com.github.mikephil.charting.components.XAxis
+import com.github.mikephil.charting.data.BarData
+import com.github.mikephil.charting.data.BarDataSet
+import com.github.mikephil.charting.data.BarEntry
+import com.github.mikephil.charting.formatter.ValueFormatter
+import com.github.mikephil.charting.interfaces.datasets.IBarDataSet
 
 class CategoryChartView @JvmOverloads constructor(
     context: Context, attrs: AttributeSet? = null, defStyleAttr: Int = 0
@@ -49,6 +56,48 @@ class CategoryChartView @JvmOverloads constructor(
         CategoryChartViewRenderer(this, this.animator, this.viewPortHandler).apply {
             mRadius = 10f
             this@CategoryChartView.renderer = this
+        }
+    }
+
+    fun setDataAndStyle(categoryList: List<String>, avgConsumptions: List<Float>) {
+        val entries = mutableListOf<BarEntry>()
+        if (categoryList.size != 5 && avgConsumptions.size != 5) {
+            return
+        }
+
+        for (index in categoryList.indices) {
+            entries.add(BarEntry(index.toFloat(), avgConsumptions[index]))
+        }
+
+        val set = BarDataSet(entries, "Bar DataSet").apply {
+            val grayColor = ContextCompat.getColor(context, R.color.colorChartGray)
+
+            val colors = listOf(
+                grayColor,
+                grayColor,
+                grayColor,
+                grayColor,
+                ContextCompat.getColor(context, R.color.colorSky)
+            )
+            setColors(colors)
+            setDrawValues(false)
+        }
+
+        val dataSets = mutableListOf<IBarDataSet>(set)
+        dataSets.add(set)
+
+        BarData(dataSets).apply {
+            this.barWidth = 0.5f
+            data = this
+        }
+
+        with(xAxis) {
+            labelCount = categoryList.size
+            valueFormatter = object : ValueFormatter() {
+                override fun getFormattedValue(value: Float): String {
+                    return categoryList[value.toInt() % categoryList.size]
+                }
+            }
         }
     }
 }
