+++ app/src/main/res/anim/mypage_category_chart_label_alpha.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<set xmlns:android="http://schemas.android.com/apk/res/android"
+    android:interpolator="@android:anim/accelerate_interpolator">
+    <alpha
+        android:duration="3000"
+        android:fromAlpha="0"
+        android:toAlpha="1.0">
+    </alpha>
+</set>
