+++ app/src/main/res/anim/mypage_lottie_move.xml
@@ -2,7 +2,7 @@
 <translate
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:interpolator="@android:anim/linear_interpolator"
-    android:fromXDelta="50"
+    android:fromXDelta="80"
     android:toXDelta="-300"
     android:fillAfter="true"
     android:duration="5000"
