+++ app/src/main/res/drawable/bg_category_chart_label.xml
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android" >
+
+    <item android:top="0dp" android:bottom="5dp">
+        <shape
+            android:shape="rectangle">
+            <solid android:color="@color/colorAccent"/>
+            <size
+                android:width="59dp"
+                android:height="15dp"/>
+            <corners android:radius="50dp" />
+        </shape>
+    </item>
+
+    <item android:bottom="0dp"
+        android:top="0dp"
+        android:gravity="center_horizontal|bottom">
+        <rotate
+            android:fromDegrees="45"
+            android:toDegrees="45"
+            android:pivotX="135%"
+            android:pivotY="15%">
+            <shape android:shape="rectangle">
+                <solid android:color="@color/colorAccent" />
+                <size
+                    android:height="5dp"
+                    android:width="5dp"/>
+            </shape>
+        </rotate>
+    </item>
+</layer-list>
