+++ app/src/main/res/drawable/bg_mypage_mycode.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android" android:shape="rectangle" >
+    <corners
+        android:radius="10dp"
+        />
+    <solid
+        android:color="@color/colorSky"
+        />
+    <padding
+        android:left="4dp"
+        android:top="2dp"
+        android:right="4dp"
+        android:bottom="2dp"
+        />
+    <size
+        android:width="40dp"
+        android:height="19dp"
+        />
+</shape>
